fix: extracted device attribute grouping into selector - this should prevent duplicating the aggregation on every state change & increase stability of the attribute list Changelog: None
