+++ src/js/components/devices/widgets/filters.js
@@ -13,6 +13,7 @@ import FilterItem from './filteritem';
 
 import { DEVICE_FILTERING_OPTIONS } from '../../../constants/deviceConstants';
 import { deepCompare } from '../../../helpers';
+import { getFilterAttributes } from '../../../selectors';
 
 export const emptyFilter = { key: null, value: '', operator: '$eq', scope: 'inventory' };
 
@@ -200,25 +201,10 @@ const actionCreators = {
 
 const mapStateToProps = (state, ownProps) => {
   const { plan = 'os' } = state.organization.organization;
-  const deviceNameAttribute = { key: 'name', value: 'Name', scope: 'tags', category: 'tags', priority: 1 };
-  const deviceIdAttribute = { key: 'id', value: 'Device ID', scope: 'identity', category: 'identity', priority: 1 };
-  const attributes = [
-    ...state.users.globalSettings.previousFilters.map(item => ({
-      ...item,
-      value: deviceIdAttribute.key === item.key ? deviceIdAttribute.value : item.key,
-      category: 'recently used',
-      priority: 0
-    })),
-    deviceNameAttribute,
-    deviceIdAttribute,
-    ...state.devices.filteringAttributes.identityAttributes.map(item => ({ key: item, value: item, scope: 'identity', category: 'identity', priority: 1 })),
-    ...state.devices.filteringAttributes.inventoryAttributes.map(item => ({ key: item, value: item, scope: 'inventory', category: 'inventory', priority: 2 })),
-    ...state.devices.filteringAttributes.tagAttributes.map(item => ({ key: item, value: item, scope: 'tags', category: 'tags', priority: 3 }))
-  ];
   const selectedGroup = state.devices.groups.selectedGroup;
   const groupFilters = state.devices.groups.byId[selectedGroup]?.filters ?? [];
   return {
-    attributes: attributes.filter((item, index, array) => array.findIndex(filter => filter.key === item.key && filter.scope === item.scope) == index),
+    attributes: getFilterAttributes(state),
     canFilterMultiple: state.app.features.isEnterprise || (state.app.features.isHosted && plan !== 'os'),
     filters: ownProps.filters || state.devices.filters || [],
     groupFilters,
